Inert tracer now used to define which cells are carried between steps
An inert tracer (denoted as `Contrail_` in the code) is defined at
the start of the 2D simulation, set to 1.0 in locations where plume
water is added. This tracer is allowed to spread, as well as being
set to 1.0 wherever contrail ice exceeds 1 particle per m3. This
is used to define locations which should be carried to the next
time step (specifically if the tracer has a value of at least 0.2).
Since teh tracer is set by ice but diffuses and advects like H2O,
it means that the influence of contrail ice is better captured.
However, this does increase run times. Users are advised to
consider changing either the ice particle number threshold or the
contrail tracer threshold (1.0 particles per m3 and 0.2,
respectively) to address this.